requirements.txt
Django==1.8.0
django-localflavor==1.1
dj-database-url==0.3.0
dj-static==0.0.6
requests==2.7.0
geocoder==1.5.0
gunicorn==19.3.0
psycopg2==2.6
pillow==2.9.0
mock==1.0.1
flake8==2.4.0
python-dateutil==2.4.2
